requirements.txt
discord.py
discord.py[voice]
ffmpeg-python
dnspython==1.16.0
PyNaCl
async-timeout==3.0.1
gTTS
python-dotenv==0.14.0
uberduckapi